feat(App): implement data seed to local storage

+++ src/utils/seedData.ts
@@ -1,4 +1,9 @@
 
 export function seedData<T>(t: T, key:string) {
-  localStorage.setItem(key, JSON.stringify(t));
+    const existing = localStorage.getItem(key);
+
+    if(!existing){
+        localStorage.setItem(key, JSON.stringify(t));
+    }
+
 }
